+++ test/utils/ArraysTest.js
@@ -177,8 +177,9 @@ suite("Arrays", function () {
     });
     test("findMatchingsProp", function () {
         var res = Arrays.findMatchingProps([{ name: "billy", value: 5 }, { name: "sam", value: 5 }, { name: "overhill", value: 3 }], "value", 5);
-        var expect = [{ name: "billy", value: 5 }, { name: "sam", value: 5 }];
-        asr.deepEqual(res, expect);
+        asr.deepEqual(res, [{ name: "billy", value: 5 }, { name: "sam", value: 5 }]);
+        var res2 = Arrays.findMatchingProps([], "value", 5);
+        asr.deepEqual(res2, []);
     });
     test("first", function () {
         var ary = [{ key: 27, value: "A" }, { key: 46, value: "B" }, { key: 84, value: "C" }, { key: 84, value: "D" }];
@@ -191,11 +192,14 @@ suite("Arrays", function () {
         asr.throws(function () { return Arrays.first(ary, function (obj) { return obj.key === 84; }, true); });
     });
     test("firstProp", function () {
-        var res2 = Arrays.firstProp([{ name: "billy", value: 5 }, { name: "sam", value: 5 }], "value", 5);
-        asr.deepEqual(res2, { name: "billy", value: 5 });
+        var res1 = Arrays.firstProp([{ name: "billy", value: 5 }, { name: "sam", value: 5 }], "value", 5);
+        asr.deepEqual(res1, { name: "billy", value: 5 });
+        var res2 = Arrays.firstProp([{ name: "billy", value: 4 }, { name: "sam", value: 5 }], "value", 4, true);
+        asr.deepEqual(res2, { name: "billy", value: 4 });
         var ary1 = [{ name: "billy", value: 4 }, { name: "sam", value: 5 }, { name: "will", value: 5 }];
-        var res1 = Arrays.firstProp(ary1, "value", 5, false);
-        asr.deepEqual(res1, { name: "sam", value: 5 });
+        var res3 = Arrays.firstProp(ary1, "value", 5, false);
+        asr.deepEqual(res3, { name: "sam", value: 5 });
+        asr.throws(function () { return Arrays.firstProp([], "value", 5, true); });
         asr.throws(function () { return Arrays.firstProp(ary1, "value", 5, true); });
     });
     test("getIfOneItem", function () {